be/int/web/captured: renames the package for cleaner selector expressions.
