+++ backend/internal/web/captured/rw.go
@@ -1,4 +1,4 @@
-package rw
+package captured
 
 import (
 	"cmp"
