+++ backend/internal/web/captured/rw_test.go
@@ -1,4 +1,4 @@
-package rw
+package captured
 
 import (
 	"bufio"
